refactor(frontend): deepen TaskRequestsRoute into queue/decision/conversion/link hooks
TaskRequestsRoute.tsx (1065 lines) mixed queue state, review-decision,
conversion, and link-existing workflows in one component. Split into
useTaskRequestsQueue plus three independent per-item workflow hooks
(useTaskRequestDecision/Conversion/Link), each exposing its own
mutation result/error shape, with TaskRequestPanel/Row/DecisionDialog
as pure presentational components. TaskRequestsRoute.tsx is now
composition only (1065 -> 78 lines).

No behavior change: all four existing test files pass unmodified
(24/24), pixel-diff against the pre-refactor render is identical
across empty, populated, and expanded-form states.

Closes #407.
